feat: execution optimization without cache limit
To render all visualizations when needed we do not have increase
execution cache limit to cover them all.
Not it supports also Edit mode and mode change.

JIRA: LX-488
risk: low
